% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_storage.R
\name{read}
\alias{read}
\title{read}
\usage{
read(name, path)
}
\arguments{
\item{name}{Character or number. Name of value.}

\item{path}{Character. Path to storage file.}
}
\value{
Character.
}
\description{
Read info from the storage file.
}
