% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_storage.R
\name{read_info}
\alias{read_info}
\title{read_info}
\usage{
read_info(name, path)
}
\arguments{
\item{name}{Character or number. Name of value.}

\item{path}{Character. Path to storage file.}
}
\value{
Character.
}
\description{
Read info from the storage file.
}
