% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_storage.R
\name{store_info}
\alias{store_info}
\title{store_info}
\usage{
store_info(name, value, path)
}
\arguments{
\item{name}{Character or number. Name of value.}

\item{value}{Character or number. Value}

\item{path}{Character. Path to storage file.}
}
\description{
Store a named value in a storage file. The storage file is created if it does not exist.
}
